+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Node `event trigger`
 
 ## [0.1.0] - 2017-06-09
 ### Added